refactor: remove unneeded ms tests
