+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/set/multiset/MultiSetImplTest.kt
@@ -19,15 +19,6 @@ internal class MultiSetImplTest {
     @Test fun testIsEmpty() = runImmutableIsEmptyTests()
     @Test fun testGetCountOf() = runImmutableGetCountOfTests()
 
-    @Test fun testMap() = runImmutableMapTests()
-    @Test fun testFilter() = runImmutableFilterTests()
-    @Test fun testFilterNot() = runImmutableFilterNotTests()
-    @Test fun testFold() = runImmutableFoldTests()
-
-    @Test fun testMapToSet() = runMapToSetTests()
-    @Test fun testFilterToSet() = runFilterToSetTests()
-    @Test fun testFilterNotToSet() = runFilterNotToSetTests()
-
     @Test
     fun testIterator() {
         var set: MultiSet<Int> = multiSetOf()
