+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/set/multiset/MutableMultiSetImplTest.kt
@@ -26,11 +26,6 @@ internal class MutableMultiSetImplTest {
     @Test fun testIsEmpty() = runMutableIsEmptyTests()
     @Test fun testGetCountOf() = runMutableGetCountOfTests()
 
-    @Test fun testMap() = runMutableMapTests()
-    @Test fun testFilter() = runMutableFilterTests()
-    @Test fun testFilterNot() = runMutableFilterNotTests()
-    @Test fun testFold() = runMutableFoldTests()
-
     @Test
     fun testIterator() {
         var set: MutableMultiSet<Int> = mutableMultiSetOf()
