+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/set/multiset/inline/runMapFoldTests.kt
@@ -1,4 +1,4 @@
-package xyz.lbres.kotlinutils.set.multiset.mutable
+package xyz.lbres.kotlinutils.set.multiset.inline
 
 import xyz.lbres.kotlinutils.list.IntList
 import xyz.lbres.kotlinutils.list.ext.copyWithoutLast
@@ -12,12 +12,12 @@ private val e1 = NullPointerException("Cannot invoke method on null value")
 private val e2 = ArithmeticException()
 private val e3 = ClassCastException("Cannot cast Int to List")
 
-internal fun runMutableMapTests() {
-    var intSet = mutableMultiSetOf<Int>()
+internal fun runMapTests() {
+    var intSet = multiSetOf<Int>()
     var expectedInt = emptyList<Int>()
     assertEquals(expectedInt, intSet.map { it * 2 }.sorted())
 
-    intSet = mutableMultiSetOf(1, 1, 5, 7, -3, 0, 2, 5)
+    intSet = multiSetOf(1, 1, 5, 7, -3, 0, 2, 5)
     expectedInt = listOf(-3, 0, 1, 1, 2, 5, 5, 7)
     assertEquals(expectedInt, intSet.map { it }.sorted())
 
@@ -27,7 +27,7 @@ internal fun runMutableMapTests() {
     var expectedString = listOf("-4", "-1", "0", "0", "1", "4", "4", "6")
     assertEquals(expectedString, intSet.map { (it - 1).toString() }.sortedBy { it.toInt() })
 
-    var stringSet = mutableMultiSetOf("hello", "world", "goodbye", "world", "hello", "hi", "world", "wrong")
+    var stringSet = multiSetOf("hello", "world", "goodbye", "world", "hello", "hi", "world", "wrong")
     val helloWorldMap: (String) -> String = {
         when (it) {
             "hello", "hi" -> "greetings"
@@ -40,7 +40,7 @@ internal fun runMutableMapTests() {
     assertEquals(expectedString.sorted(), stringSet.map { helloWorldMap(it) }.sorted())
 
     var helperString = "1"
-    stringSet = mutableMultiSetOf("1", "2", "3", "4", "5")
+    stringSet = multiSetOf("1", "2", "3", "4", "5")
     expectedString = listOf("11", "112", "1113", "11114", "111115")
     val addingMap: (String) -> String = {
         val result = "$helperString$it"
@@ -52,15 +52,15 @@ internal fun runMutableMapTests() {
     expectedString = listOf("", "", "", "", "")
     assertEquals(expectedString, stringSet.map { "" })
 
-    stringSet = mutableMultiSetOf("hello", "world", "goodbye", "world", "hello", "hi", "world", "wrong")
+    stringSet = multiSetOf("hello", "world", "goodbye", "world", "hello", "hi", "world", "wrong")
     expectedInt = listOf(1, 1, 1, 2, 2, 3, 3, 3)
     assertEquals(expectedInt, stringSet.map { stringSet.getCountOf(it) }.sorted())
 
-    val errorSet = mutableMultiSetOf(e1, e2, e3)
+    val errorSet = multiSetOf(e1, e2, e3)
     val expectedStringNull = listOf("Cannot cast Int to List", "Cannot invoke method on null value", null)
     assertEquals(expectedStringNull, errorSet.map { it.message }.sortedBy { it ?: "null" })
 
-    val listSet = mutableMultiSetOf(listOf(1, 2, 3), listOf(4, 5, 6), listOf(), listOf(7), listOf(7), listOf(7))
+    val listSet = multiSetOf(listOf(1, 2, 3), listOf(4, 5, 6), listOf(), listOf(7), listOf(7), listOf(7))
     val expectedList = listOf(listOf(), listOf(1, 2), listOf(4, 5), listOf(7), listOf(7), listOf(7))
     val listMap: (IntList) -> IntList = {
         if (it.size > 1) {
@@ -72,17 +72,17 @@ internal fun runMutableMapTests() {
     assertEquals(expectedList, listSet.map(listMap).sortedBy { if (it.isEmpty()) 0 else it.first() })
 }
 
-internal fun runMutableFoldTests() {
-    assertFailsWith<ArithmeticException> { mutableMultiSetOf(1, 2, 0).fold(1, Int::div) }
+internal fun runFoldTests() {
+    assertFailsWith<ArithmeticException> { multiSetOf(1, 2, 0).fold(1, Int::div) }
 
-    var intSet = multiSetOf<Int>()
+    var intSet = emptyMultiSet<Int>()
     var intExpected = 0
     assertEquals(intExpected, intSet.fold(0, Int::plus))
 
     intExpected = 10
     assertEquals(intExpected, intSet.fold(10, Int::plus))
 
-    intSet = mutableMultiSetOf(1, 1, 1, 2, 2, 3, 4, 5)
+    intSet = multiSetOf(1, 1, 1, 2, 2, 3, 4, 5)
     intExpected = 0
     assertEquals(intExpected, intSet.fold(0, Int::times))
 
@@ -94,20 +94,80 @@ internal fun runMutableFoldTests() {
     foldedString = foldedString.toCharArray().sorted().joinToString("")
     assertEquals(stringExpected, foldedString)
 
-    var stringSet = mutableMultiSetOf("abc", "ab", "nop", "def", "hijk", "lm", "lm", "lm", "nop")
+    var stringSet = multiSetOf("abc", "ab", "nop", "def", "hijk", "lm", "lm", "lm", "nop")
     stringExpected = "aadhlllnn"
     foldedString = stringSet.fold("") { acc, string -> string[0] + acc }
     foldedString = foldedString.toCharArray().sorted().joinToString("")
     assertEquals(stringExpected, foldedString)
 
-    stringSet = mutableMultiSetOf("abc", "abc", "de")
+    stringSet = multiSetOf("abc", "abc", "de")
     intExpected = 262144
     assertEquals(intExpected, stringSet.fold(2) { acc, string -> acc.toDouble().pow(string.length).toInt() })
 
     stringExpected = "123123123"
     assertEquals(stringExpected, stringSet.fold("") { acc, _ -> acc + "123" })
 
-    val msSet = mutableMultiSetOf(multiSetOf(1, 2, 3), multiSetOf(1, 2, 3), emptyMultiSet(), multiSetOf(9), multiSetOf(3, 3, 9, 4), emptyMultiSet())
+    val msSet = multiSetOf(multiSetOf(1, 2, 3), multiSetOf(1, 2, 3), emptyMultiSet(), multiSetOf(9), multiSetOf(3, 3, 9, 4), emptyMultiSet())
     val msExpected = multiSetOf(1, 2, 3, 1, 2, 3, 9, 3, 3, 9, 4)
     assertEquals(msExpected, msSet.fold(emptyMultiSet()) { acc, set -> acc + set })
+}
+
+internal fun runMapToSetTests() {
+    var intSet = multiSetOf<Int>()
+    var expectedInt = emptyMultiSet<Int>()
+    assertEquals(expectedInt, intSet.mapToSet { it * 2 })
+
+    intSet = multiSetOf(1, 1, 5, 7, -3, 0, 2, 5)
+    expectedInt = multiSetOf(-3, 0, 1, 1, 2, 5, 5, 7)
+    assertEquals(expectedInt, intSet.mapToSet { it })
+
+    expectedInt = multiSetOf(-6, 0, 2, 2, 4, 10, 10, 14)
+    assertEquals(expectedInt, intSet.mapToSet { it * 2 })
+
+    var expectedString = multiSetOf("-4", "-1", "0", "0", "1", "4", "4", "6")
+    assertEquals(expectedString, intSet.mapToSet { (it - 1).toString() })
+
+    var stringSet = multiSetOf("hello", "world", "goodbye", "world", "hello", "hi", "world", "wrong")
+    val helloWorldMap: (String) -> String = {
+        when (it) {
+            "hello", "hi" -> "greetings"
+            "world" -> "planet"
+            "goodbye" -> "farewell"
+            else -> "leave this planet"
+        }
+    }
+    expectedString = multiSetOf("greetings", "planet", "farewell", "planet", "greetings", "greetings", "planet", "leave this planet")
+    assertEquals(expectedString, stringSet.mapToSet { helloWorldMap(it) })
+
+    var helperString = "1"
+    stringSet = multiSetOf("1", "2", "3", "4", "5")
+    expectedString = multiSetOf("11", "112", "1113", "11114", "111115")
+    val addingMap: (String) -> String = {
+        val result = "$helperString$it"
+        helperString += "1"
+        result
+    }
+    assertEquals(expectedString, stringSet.mapToSet { addingMap(it) })
+
+    expectedString = multiSetOf("", "", "", "", "")
+    assertEquals(expectedString, stringSet.mapToSet { "" })
+
+    stringSet = multiSetOf("hello", "world", "goodbye", "world", "hello", "hi", "world", "wrong")
+    expectedInt = multiSetOf(1, 1, 1, 2, 2, 3, 3, 3)
+    assertEquals(expectedInt, stringSet.mapToSet { stringSet.getCountOf(it) })
+
+    val errorSet = multiSetOf(e1, e2, e3)
+    val expectedStringNull = multiSetOf("Cannot cast Int to List", "Cannot invoke method on null value", null)
+    assertEquals(expectedStringNull, errorSet.mapToSet { it.message })
+
+    val listSet = multiSetOf(listOf(1, 2, 3), listOf(4, 5, 6), listOf(), listOf(7), listOf(7), listOf(7))
+    val expectedList = multiSetOf(listOf(), listOf(1, 2), listOf(4, 5), listOf(7), listOf(7), listOf(7))
+    val listMap: (IntList) -> IntList = {
+        if (it.size > 1) {
+            it.copyWithoutLast()
+        } else {
+            it
+        }
+    }
+    assertEquals(expectedList, listSet.mapToSet(listMap))
 }
